requirements.txt
pyyaml
pandas
tqdm
numpy
Pillow
python-doctr
torchvision
torchaudio
ipython
wandb
transformers
timm